Some Debug Stuff, Added Enum Support and Enum DataTemplate
