+++ NoMansGUI/App.config
@@ -36,6 +36,9 @@
                     </ArrayOfString>
                 </value>
             </setting>
+            <setting name="RecentFolder" serializeAs="String">
+                <value />
+            </setting>
         </NoMansGUI.Properties.Settings>
     </userSettings>
 </configuration>
