+++ NoMansGUI/Properties/Settings.settings
@@ -26,5 +26,8 @@
   &lt;string&gt;Dark&lt;/string&gt;
 &lt;/ArrayOfString&gt;</Value>
     </Setting>
+    <Setting Name="RecentFolder" Type="System.String" Scope="User">
+      <Value Profile="(Default)" />
+    </Setting>
   </Settings>
 </SettingsFile>
